submissions/devoteam/requirements.txt
flask
sentence-transformers
PyMuPDF
python-docx
python-pptx
pandas
unidecode
faiss-cpu
tiktoken
beautifulsoup4
fpdf2
openai
python-dotenv
pdfkit